requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
croniter==1.4.1
dateutils==0.6.12
deepdiff==6.5.0
dimod==0.12.12
docker-pycreds==0.4.0
dwave-networkx==0.8.14
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
h11==0.14.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==16.0.6
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ordered-set==4.1.0
packaging==23.1
pathtools==0.1.2
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
requests==2.31.0
rich==13.5.3
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.1.2
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
wandb==0.15.10
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
yarl==1.9.2
